#2924 Move ContractChangeParameters to interface IContractChangeBL
#2924 Allow Termination of all selected contract periods
